Module kalloc::wrapper

OOM wrapper functions.

Structs

AllocError	Experimental
The AllocError error indicates an allocation failure that may be due to resource exhaustion or to something wrong when combining the given input arguments with this allocator.

HashMap
A hash map implemented with quadratic probing and SIMD lookup.

Vec
A contiguous growable array type, written Vec<T> but pronounced 'vector'.

VecDeque
A double-ended queue implemented with a growable ring buffer.

Enums

TryReserveError	Experimental
The error type for try_reserve methods.

Functions

deque_push_back
OOM Wrapper to push back an element to a deque.

deque_push_front
OOM Wrapper to push front an element to a deque.

map_insert
OOM Wrapper to insert key-valud pair to a hash map.

r2a
Map TryReserveError to AllocError for consistency.

vec_push
OOM Wrapper to push back an element into a vector. Amortized O(1).

vec_shrink_to_fit
OOM Wrapper to shrink a vector. O(N).